requirements.txt
streamlit==1.33.0
tensorflow-cpu==2.15.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-image==0.22.0